# Optimize GroupNormOp