Summary: Optimize GroupNormOp

Differential Revision: D10455567