@BIT-silence 2018-10-18
Some benchmark

On CPU, N = 128, C = 128, H = 64, W = 64, group = 32:
NCHW 214.66ms -> 152.50ms
NHWC 912.79ms -> 848.54ms

On GPU, N = 256, C = 256, H = 128, W = 128, group = 32:
NCHW 134.58ms -> 70.74ms
NHWC 221.92ms -> 102.61ms